Migrate color configuration to a popup and introduce the controller.
This is a substantial change to how this tool operates. Prior to this
commit, color configuration was done by context menus and hotkeys.

This commit introduces a new configuration dialog and a controller. The
configuration dialog is hardcoded to pop up when `m` is pressed on the
capture window. Both the controller and the configuration dialog will be
expanded upon in future commits.

The configuration window currently does not provide a means of adjusting
hsv (tricky to do right) or the pixel format (crashy). I've omitted both
primarily to keep this commit more focused.

I've removed the context menu options for configuring color since I
don't want to sync two mechanisms of configuring color. I've also
removed the hotkeys for adjusting color to prevent users from
accidentally adjusting colors.